Added new NotFoundException. Added methods to get, update and delete products. Added random generated ID's in Entity. Changed Repository from <ShopEntity, Long> to <ShopEntity, Integer>

+++ src/main/java/com/shop/repositories/ShopDatabaseRepository.java
@@ -2,5 +2,5 @@ package com.shop.repositories;
 
 import org.springframework.data.repository.CrudRepository;
 
-interface ShopDatabaseRepository extends CrudRepository<ShopEntity, Long> {
+interface ShopDatabaseRepository extends CrudRepository<ShopEntity, Integer> {
 }
